requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pyyaml
tensorboard
matplotlib
pillow
scipy
tqdm
pytorch-fid
